move generate next to other helpers
